+++ CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.62](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.61...v0.1.0-alpha.62) (2020-11-23)
+
+
+### Features
+
+* **headless:** expose hook to preprocess analytics events ([2012fe7](https://bitbucket.org/coveord/ui-kit/commits/2012fe795d74a6e21ed2c1d1c91e32ce7d19d3d0))
+
+
+
+
+
+# [0.1.0-alpha.61](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.60...v0.1.0-alpha.61) (2020-11-20)
+
+
+### Bug Fixes
+
+* **headless:** move numberOfSuggestions condition to #showSuggestions ([618d69b](https://bitbucket.org/coveord/ui-kit/commits/618d69b4fdabe7a38dbfeab218ad63d58cf57a9f))
+
+
+
+
+
+# [0.1.0-alpha.60](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.59...v0.1.0-alpha.60) (2020-11-20)
+
+
+### Bug Fixes
+
+* **headless:** make standalone search-box options optional ([1c7e4cd](https://bitbucket.org/coveord/ui-kit/commits/1c7e4cdecbc5a533a8f910d19cba1516a516235e))
+
+
+
+
+
 # [0.1.0-alpha.59](https://bitbucket.org/coveord/ui-kit/compare/v0.1.0-alpha.58...v0.1.0-alpha.59) (2020-11-20)